+++ test/unit/math/prim/prob/beta_binomial_log_test.cpp
@@ -14,18 +14,17 @@ TEST(ProbBetaBinomial, log_matches_lpmf) {
   EXPECT_FLOAT_EQ((stan::math::beta_binomial_lpmf<false>(n, N, alpha, beta)),
                   (stan::math::beta_binomial_log<false>(n, N, alpha, beta)));
   EXPECT_FLOAT_EQ(
-      (stan::math::beta_binomial_lpmf<true, double, double, double, double>(
+      (stan::math::beta_binomial_lpmf<true, int, int, double, double>(
           n, N, alpha, beta)),
-      (stan::math::beta_binomial_log<true, double, double, double, double>(
+      (stan::math::beta_binomial_log<true, int, int, double, double>(
           n, N, alpha, beta)));
   EXPECT_FLOAT_EQ(
-      (stan::math::beta_binomial_lpmf<false, double, double, double, double>(
+      (stan::math::beta_binomial_lpmf<false, int, int, double, double>(
           n, N, alpha, beta)),
-      (stan::math::beta_binomial_log<false, double, double, double, double>(
+      (stan::math::beta_binomial_log<false, int, int, double, double>(
           n, N, alpha, beta)));
-  EXPECT_FLOAT_EQ(
-      (stan::math::beta_binomial_lpmf<double, double, double, double>(
-          n, N, alpha, beta)),
-      (stan::math::beta_binomial_log<double, double, double, double>(
-          n, N, alpha, beta)));
+  EXPECT_FLOAT_EQ((stan::math::beta_binomial_lpmf<int, int, double, double>(
+                      n, N, alpha, beta)),
+                  (stan::math::beta_binomial_log<int, int, double, double>(
+                      n, N, alpha, beta)));
 }
