+++ test/unit/math/prim/prob/binomial_log_test.cpp
@@ -12,10 +12,10 @@ TEST(ProbBinomial, log_matches_lpmf) {
                   (stan::math::binomial_log<true>(n, N, theta)));
   EXPECT_FLOAT_EQ((stan::math::binomial_lpmf<false>(n, N, theta)),
                   (stan::math::binomial_log<false>(n, N, theta)));
-  EXPECT_FLOAT_EQ((stan::math::binomial_lpmf<true, double>(n, N, theta)),
-                  (stan::math::binomial_log<true, double>(n, N, theta)));
-  EXPECT_FLOAT_EQ((stan::math::binomial_lpmf<false, double>(n, N, theta)),
-                  (stan::math::binomial_log<false, double>(n, N, theta)));
-  EXPECT_FLOAT_EQ((stan::math::binomial_lpmf<double>(n, N, theta)),
-                  (stan::math::binomial_log<double>(n, N, theta)));
+  EXPECT_FLOAT_EQ((stan::math::binomial_lpmf<true, int>(n, N, theta)),
+                  (stan::math::binomial_log<true, int>(n, N, theta)));
+  EXPECT_FLOAT_EQ((stan::math::binomial_lpmf<false, int>(n, N, theta)),
+                  (stan::math::binomial_log<false, int>(n, N, theta)));
+  EXPECT_FLOAT_EQ((stan::math::binomial_lpmf<int>(n, N, theta)),
+                  (stan::math::binomial_log<int>(n, N, theta)));
 }
